requirements.txt
fastapi==0.115.7
pydantic==2.10.6
pydantic-settings==2.7.1
openai==1.60.0
PyJWT==2.10.1
requests==2.32.3
uvicorn==0.34.0
redis==5.2.1
cachetools==5.5.2
python-frontmatter
markdown

# 如有数据库可加：
# sqlalchemy
# asyncpg
# aiomysql
# 其他依赖根据实际项目补充
